FlatVel_A/requirements.txt
black>=24.1.0
numpy>=3.8.2
omegaconf>=2.3.0
torch>=2.1.2
matplotlib>=3.8.2
hydra-core>=1.3.2
